changed folder structures, got login to work, and modified flash messages for user feedback
