+++ JDID/classes/__init__.py
@@ -1,0 +1,4 @@
+#!/usr/bin/python3
+from JDID.database.engine import DBStorage
+
+storage = DBStorage()
