+++ JDID/classes/goal_class.py
