+++ JDID/templates/login.html
@@ -5,29 +5,31 @@
         {{ form.hidden_tag() }}
         <div class="col align-self-center">
             <div>
-                {{ form.email.label(class="form-control") }}
+                {{ form.email.label(class="form-control form-control-sm") }}
                 <!-- error -->
                 {% if form.email.errors %}
-                <div class='invalid-input'>
+                {{ form.email(class="form-control-sm is-invalid")}}
+                <div class='invalid-feedback'>
                     {% for error in form.email.errors %}
                     <span>{{ error }}</span>
                     {% endfor %}
                 </div>
                 {% else %}
-                {{ form.email(class="form-control") }}
+                {{ form.email(class="form-control form-control-sm") }}
                 {% endif %}
             </div>
             <div>
-                {{ form.password.label(class="form-control") }}
+                {{ form.password.label(class="form-control form-control-sm") }}
                 <!-- error -->
                 {% if form.password.errors %}
-                <div class='invalid-input'>
+                {{ form.password(class="form-control-sm is-invalid")}}
+                <div class='invalid-feedback'>
                     {% for error in form.password.errors %}
                     <span>{{ error }}</span>
                     {% endfor %}
                 </div>
                 {% else %}
-                {{ form.password(class="form-control") }}
+                {{ form.password(class="form-control form-control-sm") }}
                 {% endif %}
             </div>
         </div>
