--- classes/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/python3
-
-
-from database.engine import DBStorage
-
-
-storage = DBStorage()
